countinue WPF

+++ PL/BusWindow.xaml.cs
@@ -37,14 +37,10 @@ namespace PL
             InitializeComponent();
             bl = _bl;
             refreshListBus();
-            detailOfBus.Visibility = Visibility.Hidden;
-            refuel.Visibility = Visibility.Hidden;
-            treat.Visibility = Visibility.Hidden;
-            delete.Visibility = Visibility.Hidden;
-            update.Visibility = Visibility.Hidden;
+            
         }
 
-        private void listBus_MouseDoubleClick(object sender, MouseButtonEventArgs e)
+        /*private void listBus_MouseDoubleClick(object sender, MouseButtonEventArgs e)
         {
             BO.Bus bus = listBus.SelectedItem as BO.Bus;
             frame.Visibility = Visibility.Hidden;
@@ -59,10 +55,12 @@ namespace PL
             sumKm.Text = "" + bus.sumKM;
             fuel.Text = "" + bus.totalFuel;
             KmFrom.Text = "" + bus.sumKMFromLastTreat;
-        }
+            status.Text = "" + bus.Status;
+        }*/
 
-        private void update_Click(object sender, RoutedEventArgs e)
+        /*private void update_Click(object sender, RoutedEventArgs e)
         {
+
             BO.Bus temp = new BO.Bus();
             temp.numberLicense = (string)numberBus.Text;
             temp.startActivity = (DateTime)dateBus.SelectedDate;
@@ -97,39 +95,81 @@ namespace PL
                 MessageBox.Show(ex.Message, "ERROR", MessageBoxButton.OK, MessageBoxImage.Error);
             }
 
-        }
+        }*/
 
-        private void delete_Click(object sender, RoutedEventArgs e)
+        /*private void delete_Click(object sender, RoutedEventArgs e)
         {
             BO.Bus bus = listBus.SelectedItem as BO.Bus;
-            bl.deleteBus(bus);
-            refreshListBus();
-        }
+            MessageBoxResult delete = MessageBox.Show($"did you shure to delete bus {bus}?", "delete bus", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (delete == MessageBoxResult.Yes)
+            {
+                try
+                {
+                    bl.deleteBus(bus);
+                    refreshListBus();
+                    MessageBox.Show("delete complete", "OK", MessageBoxButton.OK, MessageBoxImage.None);
+                }
+                catch (BO.BadIDExceptions ex)
+                {
+                    MessageBox.Show(ex.Message, "ERROR", MessageBoxButton.OK, MessageBoxImage.Error);
+                }
+            }
+
+        }*/
 
         private void DoubleClickToAddNewBus(object sender, MouseButtonEventArgs e)
         {
-            detailOfBus.Visibility = Visibility.Hidden;
+            /*detailOfBus.Visibility = Visibility.Hidden;
+            refuel.Visibility = Visibility.Hidden;
+            treat.Visibility = Visibility.Hidden;
+            delete.Visibility = Visibility.Hidden;
+            update.Visibility = Visibility.Hidden;*/
             frame.Visibility = Visibility.Visible;
             frame.Content = new AddNewBus(bl, listBus);
         }
 
-        private void refuel_Click(object sender, RoutedEventArgs e)
+        /*private void refuel_Click(object sender, RoutedEventArgs e)
         {
             BO.Bus bus = listBus.SelectedItem as BO.Bus;
-            bl.refuel(bus);
-            refreshListBus();
-        }
+            try
+            {
+                bl.refuel(bus);
+                fuel.Text = "" + bus.totalFuel;
+                refreshListBus();
+            }
+            catch(BO.BadIDExceptions ex)
+            {
+                MessageBox.Show(ex.Message, "ERROR", MessageBoxButton.OK, MessageBoxImage.Error);
+            }
 
-        private void treat_Click(object sender, RoutedEventArgs e)
+        }*/
+
+        /*private void treat_Click(object sender, RoutedEventArgs e)
         {
             BO.Bus bus = listBus.SelectedItem as BO.Bus;
-            bl.treat(bus);
-            refreshListBus();
-        }
+            try
+            {
+                bl.treat(bus);
+                refreshListBus();
+                dateTreat.DataContext = bus;
+                KmFrom.Text = "" + bus.sumKMFromLastTreat;
+            }
+            catch (BO.BadIDExceptions ex)
+            {
+                MessageBox.Show(ex.Message, "ERROR", MessageBoxButton.OK, MessageBoxImage.Error);
+            }
+        }*/
 
         private void numberBus_PreviewTextInput(object sender, TextCompositionEventArgs e)
         {
             e.Handled = !e.Text.Any(x => char.IsDigit(x));
+        }
+
+        private void listBus_MouseDoubleClick(object sender, MouseButtonEventArgs e)
+        {
+            BO.Bus bus = listBus.SelectedItem as BO.Bus;
+            frame.Visibility = Visibility.Visible;
+            frame.Content = new DetailOfBus(bl, bus, listBus);
         }
     }
 }
